+++ tests/testthat/rnaseq.R
@@ -1,5 +1,6 @@
 library(bcbioR)
 
+
 test_that("rnaseq testing", {
   path <- withr::local_tempdir()
   print(path)