+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fix when sample start by number
 * Fix when rRNA rate is missing
 * Add by sample plots in QC
+* Add function to check nfcore samplesheet
 
 # bcbioR 0.1.1